スクレイピング処理をクラスに切り出した
